more securitygit add *git add *

+++ src/phazeid/main.ts
@@ -8,13 +8,11 @@ import { SignupRequestBody, SignupRequestBodyType } from "./types/SignupRequestB
 import { VerifyRequestBody, VerifyRequestBodyType } from "./types/VerifyRequestBody";
 import { LoginRequestBody, LoginRequestBodyType } from "./types/LoginRequestBody";
 import { ResponseError } from "./types/ResponseError";
-import { SignupResponse } from "./types/SignupResponse";
-import { LoginResponse } from "./types/LoginResponse";
 
 import * as aviUtils from "./aviUtils";
 import users from "./db/users";
 import sessions from "./db/sessions";
-import { VerifyResponse } from "./types/VerifyResponse";
+import { cleanSessionsForUser } from "./sessionUtils";
 
 let main = ( fastify: FastifyInstance ) => {
   mongoose.connect(process.env.MONGO_URI!)
@@ -41,7 +39,8 @@ let main = ( fastify: FastifyInstance ) => {
           400: ResponseError,
           409: ResponseError,
           406: ResponseError,
-          200: SignupResponse
+          500: ResponseError,
+          200: { ok: { type: 'boolean' }, session: { type: 'string' } }
         }
       }
     },
@@ -58,10 +57,10 @@ let main = ( fastify: FastifyInstance ) => {
       user = await users.findOne({ email: req.body.email });
       if(user)return reply.code(409).send({ ok: false, error: 'User Already Exists' });
 
-      let ipReq = await fetch(`https://ipinfo.io/${req.headers['cf-connecting-ip']}?token=96a00067d1963b`);
+      let ipReq = await fetch(`https://ipinfo.io/${req.headers['cf-connecting-ip']}?token=${process.env.IPINFO_KEY}`);
       let ipInfo = await ipReq.json();
 
-      let userID = crypto.randomUUID()
+      let userID = crypto.randomUUID();
 
       let session = await sessions.create({
         _id: crypto.randomUUID(),
@@ -69,6 +68,8 @@ let main = ( fastify: FastifyInstance ) => {
         createdOn: new Date(),
         expiresOn: new Date(Date.now() + 259200000),
         loc: ipInfo,
+        valid: true,
+        challengeCode: '',
         userID
       })
 
@@ -110,7 +111,7 @@ let main = ( fastify: FastifyInstance ) => {
 
       if(sent.err){
         console.error(sent.err);
-        return reply.send({ ok: false, error: 'Failed to verify email' });
+        return reply.code(500).send({ ok: false, error: 'Failed to verify email' });
       }
 
       aviUtils.generateAvi(userData.username, userData._id + '/' + userData.avatar);
@@ -128,12 +129,13 @@ let main = ( fastify: FastifyInstance ) => {
         response: {
           400: ResponseError,
           403: ResponseError,
-          200: LoginResponse
+          200: { ok: { type: 'boolean' }, session: { type: 'string' }, requiresMfa: { type: 'boolean' }, valid: { type: 'boolean' } }
         }
       }
     },
     async ( req, reply ) => {
       reply.header('Content-Type', 'application/json');
+      if(!req.headers['cf-connecting-ip'])return reply.code(400).send({ ok: false, error: 'Invalid Request Body' });
 
       if(req.headers["content-type"] !== 'application/json')return reply.code(400).send({ ok: false, error: 'Invalid Request Body' });
       if(!req.body || !req.body.username || !req.body.password)return reply.code(400).send({ ok: false, error: 'Invalid Request Body' });
@@ -143,10 +145,243 @@ let main = ( fastify: FastifyInstance ) => {
 
       let isValid = await argon2.verify(user.password!, req.body.password.toString(), { type: argon2.argon2id });
 
-      if(isValid)
-        reply.send({ ok: true, session: 'no', requiresMfa: user.hasMfa })
-      else
+      if(isValid){
+        let sessionsList = await cleanSessionsForUser(user._id!);
+
+        let ipReq = await fetch(`https://ipinfo.io/${req.headers['cf-connecting-ip']}?token=${process.env.IPINFO_KEY}`);
+        let ipInfo = await ipReq.json();
+
+        let sessionValid = sessionsList.find(x => x.loc.ip === ipInfo.ip);
+        if(!sessionValid || !sessionValid.valid)sessionValid = null;
+
+        let session = await sessions.create({
+          _id: crypto.randomUUID(),
+          token: crypto.randomBytes(32).toString('hex'),
+          createdOn: new Date(),
+          expiresOn: new Date(Date.now() + 259200000),
+          loc: ipInfo,
+          valid: sessionValid ? true : false,
+          challengeCode: sessionValid ? '' : Math.floor(Math.random() * 1_000_000).toString().padStart(6, '0'),
+          userID: user._id
+        })
+
+        if(!session.valid){
+          let mail = () => {
+            return new Promise<{ err: any, info: any }>(( res, rej ) => {
+              transport.sendMail({
+                from: 'Phaze ID <no-reply@phazed.xyz>',
+                to: user!.email!,
+                subject: 'Verification Email',
+                html: `Your verication code is ${ session.challengeCode }<br />Do <b>NOT</b> share this code with anyone.`
+              }, ( err, info ) => {
+                res({ err, info });
+              })
+            })
+          }
+    
+          let sent = await mail();
+
+          if(sent.err){
+            console.error(sent.err);
+
+            await sessions.deleteOne({ _id: session._id });
+            return reply.send({ ok: false, error: 'Failed to verify email' });
+          }
+        }
+
+        user.sessions.push(session._id!);
+        await user.save();
+
+        reply.send({ ok: true, session: session.token, requiresMfa: user.hasMfa, valid: session.valid })
+      } else
         reply.code(403).send({ ok: false, error: 'Incorrect Username or Password' });
+    }
+  )
+
+  fastify.get<{ Querystring: { token: String } }>(
+    '/api/id/v1/auth/sessions',
+    {
+      schema: {
+        response: {
+          400: ResponseError,
+          401: ResponseError,
+          403: ResponseError,
+          200: { 
+            ok: { type: 'boolean' },
+            currentSession: { type: 'string' },
+            sessionCount: { type: 'number' },
+            sessions: {
+              type: 'array',
+              items: {
+                type: 'object',
+                properties: {
+                  _id: { type: 'string' },
+                  createdOn: { type: 'number' },
+                  expiresOn: { type: 'number' },
+                  valid: { type: 'boolean' },
+                  loc: {
+                    type: 'object',
+                    properties: {
+                      ip: { type: 'string' },
+                      hostname: { type: 'string' },
+                      city: { type: 'string' },
+                      region: { type: 'string' },
+                      country: { type: 'string' },
+                      loc: { type: 'string' },
+                      org: { type: 'string' },
+                      postal: { type: 'string' },
+                      timezone: { type: 'string' }
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    },
+    async ( req, reply ) => {
+      reply.header('Content-Type', 'application/json');
+
+      if(!req.headers['cf-connecting-ip'])return reply.code(400).send({ ok: false, error: 'Invalid Request' });
+      if(!req.query.token)return reply.code(400).send({ ok: false, error: 'Invalid Query String' });
+
+      let session = await sessions.findOne({ token: req.query.token });
+      if(!session)return reply.code(401).send({ ok: false, error: 'Invalid Token' });
+
+      if(req.headers['cf-connecting-ip'] !== session.loc!.ip)
+        return reply.code(401).send({ ok: false, error: 'Invalid Session' });
+      
+      let user = await users.findById(session.userID);
+      if(!user){
+        sessions.deleteOne({ _id: session._id });
+        return reply.code(401).send({ ok: false, error: 'Invalid Session' });
+      }
+
+      if(!session.expiresOn || session.expiresOn.getTime() < Date.now()){
+        user.sessions = user.sessions.filter(x => x !== session!._id);
+        await user.save();
+
+        sessions.deleteOne({ _id: session._id });
+        return reply.code(401).send({ ok: false, error: 'Invalid Session' });
+      }
+
+      if(!session.valid)
+        return reply.code(403).send({ ok: false, error: 'Session requires verification' });
+
+      let sessionsList = await cleanSessionsForUser(user._id!);
+      reply.send({ ok: true, currentSession: session._id, sessionCount: sessionsList.length, sessions: sessionsList.map(x => { return { _id: x._id, valid: x.valid, createdOn: x.createdOn.getTime(), expiresOn: x.expiresOn.getTime(), loc: x.loc } }) })
+    }
+  )
+
+  fastify.delete<{ Querystring: { sessionId: string, token: String } }>(
+    '/api/id/v1/auth/sessions',
+    {
+      schema: {
+        response: {
+          400: ResponseError,
+          401: ResponseError,
+          404: ResponseError,
+          403: ResponseError,
+          200: { ok: { type: 'boolean' } }
+        }
+      }
+    },
+    async ( req, reply ) => {
+      reply.header('Content-Type', 'application/json');
+
+      if(!req.headers['cf-connecting-ip'])return reply.code(400).send({ ok: false, error: 'Invalid Request' });
+      if(!req.query.token)return reply.code(400).send({ ok: false, error: 'Invalid Query String' });
+
+      let session = await sessions.findOne({ token: req.query.token });
+      if(!session)return reply.code(401).send({ ok: false, error: 'Invalid Token' });
+
+      if(req.headers['cf-connecting-ip'] !== session.loc!.ip)
+        return reply.code(401).send({ ok: false, error: 'Invalid Session' });
+      
+      let user = await users.findById(session.userID);
+      if(!user){
+        sessions.deleteOne({ _id: session._id });
+        return reply.code(401).send({ ok: false, error: 'Invalid Session' });
+      }
+
+      if(!session.expiresOn || session.expiresOn.getTime() < Date.now()){
+        user.sessions = user.sessions.filter(x => x !== session!._id);
+        await user.save();
+
+        sessions.deleteOne({ _id: session._id });
+        return reply.code(401).send({ ok: false, error: 'Invalid Session' });
+      }
+
+      if(!session.valid)
+        return reply.code(403).send({ ok: false, error: 'Session requires verification' });
+
+      let sessionToRemove = await sessions.findById(req.query.sessionId);
+      if(!sessionToRemove)return reply.code(404).send({ ok: false, error: 'Cannot find session' });
+
+      user.sessions = user.sessions.filter(x => x !== req.query.sessionId);
+
+      await sessions.deleteOne({ _id: req.query.sessionId });
+      await user.save();
+
+      reply.send({ ok: true });
+    }
+  )
+
+  fastify.post<{ Body: VerifyRequestBodyType, Querystring: { token: String } }>(
+    '/api/id/v1/auth/sessions/verify',
+    {
+      schema: {
+        body: VerifyRequestBody,
+        response: {
+          400: ResponseError,
+          401: ResponseError,
+          403: ResponseError,
+          409: ResponseError,
+          200: { ok: { type: 'boolean' } }
+        }
+      }
+    },
+    async ( req, reply ) => {
+      reply.header('Content-Type', 'application/json');
+      if(!req.headers['cf-connecting-ip'])return reply.code(400).send({ ok: false, error: 'Invalid Request' });
+
+      if(req.headers["content-type"] !== 'application/json')return reply.code(400).send({ ok: false, error: 'Invalid Request Body' });
+      if(!req.body || !req.body.code)return reply.code(400).send({ ok: false, error: 'Invalid Request Body' });
+
+      if(!req.query.token)return reply.code(400).send({ ok: false, error: 'Invalid Query String' });
+
+      let session = await sessions.findOne({ token: req.query.token });
+      if(!session)return reply.code(401).send({ ok: false, error: 'Invalid Token' });
+
+      if(req.headers['cf-connecting-ip'] !== session.loc!.ip)
+        return reply.code(401).send({ ok: false, error: 'Invalid Session' });
+      
+      let user = await users.findById(session.userID);
+      if(!user){
+        sessions.deleteOne({ _id: session._id });
+        return reply.code(401).send({ ok: false, error: 'Invalid Session' });
+      }
+
+      if(!session.expiresOn || session.expiresOn.getTime() < Date.now()){
+        user.sessions = user.sessions.filter(x => x !== session!._id);
+        await user.save();
+
+        sessions.deleteOne({ _id: session._id });
+        return reply.code(401).send({ ok: false, error: 'Invalid Session' });
+      }
+
+      if(session.valid)
+        return reply.code(409).send({ ok: false, error: 'Email already verified' });
+
+      if(session.challengeCode !== req.body.code)
+        return reply.code(403).send({ ok: false, error: 'Invalid Code' });
+
+      session.valid = true;
+      session.challengeCode = '';
+      await session.save();
+
+      reply.send({ ok: true });
     }
   )
 
@@ -160,12 +395,14 @@ let main = ( fastify: FastifyInstance ) => {
           400: ResponseError,
           401: ResponseError,
           403: ResponseError,
-          200: VerifyResponse
+          409: ResponseError,
+          200: { ok: { type: 'boolean' } }
         }
       }
     },
     async ( req, reply ) => {
       reply.header('Content-Type', 'application/json');
+      if(!req.headers['cf-connecting-ip'])return reply.code(400).send({ ok: false, error: 'Invalid Request' });
 
       if(req.headers["content-type"] !== 'application/json')return reply.code(400).send({ ok: false, error: 'Invalid Request Body' });
       if(!req.body || !req.body.code)return reply.code(400).send({ ok: false, error: 'Invalid Request Body' });
@@ -175,14 +412,28 @@ let main = ( fastify: FastifyInstance ) => {
       let session = await sessions.findOne({ token: req.query.token });
       if(!session)return reply.code(401).send({ ok: false, error: 'Invalid Token' });
 
-      // if(!session.expiresOn || session.expiresOn.getTime() > Date.now())
-      //   return reply.code(401).send({ ok: false, error: 'Invalid Session' });
+      if(req.headers['cf-connecting-ip'] !== session.loc!.ip)
+        return reply.code(401).send({ ok: false, error: 'Invalid Session' });
       
       let user = await users.findById(session.userID);
       if(!user){
-        sessions.deleteOne({ _id: session._id })
+        sessions.deleteOne({ _id: session._id });
         return reply.code(401).send({ ok: false, error: 'Invalid Session' });
       }
+
+      if(!session.expiresOn || session.expiresOn.getTime() < Date.now()){
+        user.sessions = user.sessions.filter(x => x !== session!._id);
+        await user.save();
+
+        sessions.deleteOne({ _id: session._id });
+        return reply.code(401).send({ ok: false, error: 'Invalid Session' });
+      }
+
+      if(!session.valid)
+        return reply.code(403).send({ ok: false, error: 'Session requires verification' });
+
+      if(user.emailVerified)
+        return reply.code(409).send({ ok: false, error: 'Email already verified' });
 
       if(user.emailVerificationCode !== req.body.code)
         return reply.code(403).send({ ok: false, error: 'Invalid Code' });
